add loyalty program settings

+++ src/main/java/ru/hse/diplom/cafe_recommend_backend/model/dto/UserDto.java
@@ -13,6 +13,6 @@ public class UserDto {
     private String firstName;
     private String lastName;
     private String phone;
-    private Integer loyaltyDiscount;
+    private Double loyaltyDiscount;
     private OffsetDateTime createdAt;
 }
